+++ editor/src/routes/editor.svelte
@@ -8,7 +8,7 @@
     let editorContainer: HTMLElement;
     let wrapper: MonacoEditorLanguageClientWrapper | undefined;
     let markers: editor.IMarker[] = $state([]);
-    let content = $state('SELECT * WHERE {\n  ?s ?p ?o\n}');
+    let content = $state('SELECT * WHERE {\n  ?s <http://www.w3.org/2000/01/rdf-schema#> ?o .\n}');
     let cursorOffset = $state(0);
 
     onMount(async () => {